Remove redundant array_filter() calls
PHPStan v2 is reporting that \PhpParser\Node\Stmt\Interface_->extends cannot
contain falsy values. The PHPDoc type for this property declares its type as
`\PhpParser\Node\Name[]`. Therefore, `[null]` is not a valid value for this
property. Taking this into account, calling array_filter() on the property is
not necessary.

Additionally, the code for \PhpParser\Node\Stmt\Class_->extends can be
simplified to remove the array_map() and array_filter() calls.

As a bonus, one violation in the Psalm baseline has been removed as part of
this change.